requirements.txt
# Telegram: официальный клиент Bot API (async). Пакет telebot/pyTelegramBotAPI не нужен.
python-telegram-bot>=21.0
ccxt>=4.0
pandas>=2.0
numpy>=1.24
APScheduler>=3.10
python-dotenv>=1.0
ta>=0.11.0
matplotlib>=3.8.0
